Move transport errors to transport::error module
This cleans up the transport module by moving error definitions to
a separate transport::error module. This follows our normal pattern of
defining errors in a separate file.

The errors are re-exported from transport with pub use, so this is not
a functional or API change.
